Use refactored ZK & broken-out Exhibitor cookbooks
Keeping up with the Joneses

Some tests have been dropped, as the Exhibitor service script
it now handled using runit, instead of Upstart.

Exhibitor attributes are broken out into a separate file.

The exhibitor cookbook has to be brought in from GitHub, as its
not published to Supermarket yet (ride the bullet)